# Streamline proto serialization of circuits with duplicate operations