- Some circuits, such as randomized benchmarking circuits, have many repeated operations.
- Streamline the proto serialization of these circuits so that these operations are stored only once (in the constants table).

Note that this PR only adds deserialization by default and hides serialization behind a feature flag. Once this is deployed everywhere, then we can enable serialization.

Example serialization of a randomize benchmarking circuit with 70 qubits and 500 cliffords:

Before change:
- Serialization time;  0.48 s, Deserialization time: 1.08 s
- Serialization size: 2129715

After change:
- Serialization time;  0.13 s, Deserialization time: 0.10 s
- Serialization size: 174852
-